Update LICENSE
The copyright of the license mentions Nickel Lang, but this entity doesn't really have any legal existence, nor clear definition. It's mostly a GitHub org, for the time being. This PR sets the copyright owner to the company holding Tweag, which is Modus Create LLC, in line with other Tweag repositories.